+++ src/linear_interpolator.cpp
@@ -1,7 +1,7 @@
 #include "linear_interpolator.h"
 
 
-LinearInterpolator::LinearInterpolator(float input_min, float input_max, float output_min, float output_max):
+LinearInterpolator::LinearInterpolator(double input_min, double input_max, double output_min, double output_max):
     x1(input_min),
     x2(input_max),
     y1(output_min),
@@ -13,9 +13,9 @@ LinearInterpolator::~LinearInterpolator() {
 
 }
 
-float LinearInterpolator::map(float value) const {
-    float a = (y2 - y1) / (x2 - x1);
-    float b = value - x1;
+double LinearInterpolator::map(double value) const {
+    double a = (y2 - y1) / (x2 - x1);
+    double b = value - x1;
 
     return y1 + a * b;
 }
